+++ tests/test_model.py
@@ -377,6 +377,16 @@ async def test_process_event(qtbot: QtBot, model: Model) -> None:
             message={"id": "innerLoopControlMode", "address": 1, "mode": 2}
         )
 
+    with qtbot.waitSignal(
+        model.fault_manager.signal_error.summary_faults_status, timeout=TIMEOUT
+    ):
+        await model._process_event(message={"id": "summaryFaultsStatus", "status": 10})
+
+    with qtbot.waitSignal(
+        model.fault_manager.signal_error.enabled_faults_mask, timeout=TIMEOUT
+    ):
+        await model._process_event(message={"id": "enabledFaultsMask", "mask": 10})
+
 
 def test_get_message_name(model: Model) -> None:
     assert model._get_message_name("") == ""